[FLINK-13143] Refactor CheckpointExceptionHandler relevant classes
